Only route pages when the full request path is consumed
This is a breaking change for applications that depend on the
previous behavior which was similar to having an implicit
CaptureAll in front of every View.

After this change, servant-router more accurately reflects the
behavior of servant-server. Another benefit is that the ordering
of endpoints in the API becomes less relevant.